+++ NEWS.md
@@ -1,4 +1,13 @@
-# semlbci 0.10.3.5
+# semlbci 0.10.4
+
+## New Feature
+
+- The print method of the output of
+  `semlbci()` (`print.semlbci()`) can
+  print the results in a `lavaan`-like
+  format. (0.10.4).
+
+## Miscellaneous
 
 - Updated README.md for CRAN release. Identical
   to the CRAN release in code. (0.10.3.1)
@@ -7,6 +16,7 @@
   for a coming changes to the parser of
   `lavaan` (0.10.3.3).
 - Updated a few tests (0.10.3.4, 0.10.3.5).
+- Finalized to 0.10.4.
 
 # semlbci 0.10.3